+++ src/textual/pilot.py
@@ -100,7 +100,7 @@ class Pilot(Generic[ReturnType]):
         app = self.app
         screen = app.screen
         if selector is not None:
-            target_widget = screen.query_one(selector)
+            target_widget = app.query_one(selector)
         else:
             target_widget = screen
 
@@ -132,7 +132,7 @@ class Pilot(Generic[ReturnType]):
         app = self.app
         screen = app.screen
         if selector is not None:
-            target_widget = screen.query_one(selector)
+            target_widget = app.query_one(selector)
         else:
             target_widget = screen