+++ tests/test_pilot.py
@@ -52,3 +52,21 @@ async def test_pilot_exception_catching_action():
     with pytest.raises(ZeroDivisionError):
         async with FailingApp().run_test() as pilot:
             await pilot.press("b")
+
+
+async def test_pilot_click_screen():
+    """Regression test for https://github.com/Textualize/textual/issues/3395.
+
+    Check we can use `Screen` as a selector for a click."""
+
+    async with App().run_test() as pilot:
+        await pilot.click("Screen")
+
+
+async def test_pilot_hover_screen():
+    """Regression test for https://github.com/Textualize/textual/issues/3395.
+
+    Check we can use `Screen` as a selector for a hover."""
+
+    async with App().run_test() as pilot:
+        await pilot.hover("Screen")